refactor: Main view 에서 모듈 분리
